Add Application.getMashups(), check for mashups dependent on application before trying to delete it

+++ viewer-admin/src/main/java/nl/b3p/viewer/admin/stripes/ChooseApplicationActionBean.java
@@ -71,9 +71,9 @@ public class ChooseApplicationActionBean extends ApplicationActionBean {
     private Application applicationWorkversion;
     @Validate
     private Application applicationToDelete;
-    
+
     private List<Application> apps;
-    
+
     private String defaultAppId;
 
     @Validate
@@ -184,7 +184,7 @@ public class ChooseApplicationActionBean extends ApplicationActionBean {
         this.defaultApplication = defaultApplication;
     }
     //</editor-fold>
-    
+
     @DefaultHandler
     public Resolution view() {
         return new ForwardResolution(JSP);
@@ -211,10 +211,21 @@ public class ChooseApplicationActionBean extends ApplicationActionBean {
                 applicationToDelete.setVersion(uniqueVersion);
                 Stripersist.getEntityManager().getTransaction().commit();
             } else {
-                Stripersist.getEntityManager().remove(applicationToDelete);
-                Stripersist.getEntityManager().getTransaction().commit();
+                List<Application> mashups = applicationToDelete.getMashups();
+                if(!mashups.isEmpty()) {
+                    List<String> list = new ArrayList();
+                    for(Application mashup: mashups) {
+                        list.add(mashup.getNameWithVersion());
+                    }
+                    String mashupList = String.join(", ", list);
+                    getContext().getValidationErrors().addGlobalError(new SimpleError("Deze applicatie kan niet verwijderd worden, omdat de boomstructuur wordt gebruikt in de mashups " + mashupList));
+                } else {
 
-                getContext().getMessages().add(new SimpleMessage("Applicatie is verwijderd"));
+                    Stripersist.getEntityManager().remove(applicationToDelete);
+                    Stripersist.getEntityManager().getTransaction().commit();
+
+                    getContext().getMessages().add(new SimpleMessage("Applicatie is verwijderd"));
+                }
             }
             if (applicationToDelete.equals(application)) {
                 setApplication(null);
@@ -401,7 +412,7 @@ public class ChooseApplicationActionBean extends ApplicationActionBean {
         em.getTransaction().commit();
         return copy;
     }
-    
+
     public Resolution saveDefaultApplication() throws JSONException {
         JSONObject json = new JSONObject();
 
@@ -429,7 +440,7 @@ public class ChooseApplicationActionBean extends ApplicationActionBean {
         }
         return new StreamingResolution("application/json", new StringReader(json.toString()));
     }
-    
+
     @After(stages = {LifecycleStage.BindingAndValidation})
     public void createLists() {
         EntityManager em = Stripersist.getEntityManager();
